add logic for view result. Add script for converter alart(resultBox)

+++ Currency converter C#/Currency converter/Services/IConverterServices.cs
@@ -2,6 +2,6 @@
 {
     public interface IConverterServices
     {
-        string Converter(int inputAmount, string formControl, string toControl);
+        string Converter(double inputAmount, string formControl, string toControl);
     }
 }
